requirements.txt
pytesseract
Pillow
keyboard
requests
pyautogui
plyer
